+++ src/Areas/Search/Services/ISearchService.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using AzureMcp.Areas.Search.Models;
 using AzureMcp.Options;
 using static AzureMcp.Areas.Search.Commands.Index.IndexDescribeCommand;
 
@@ -13,30 +14,15 @@ public interface ISearchService
         string? tenantId = null,
         RetryPolicyOptions? retryPolicy = null);
 
-    Task<List<string>> ListIndexes(
+    Task<List<IndexInfo>> ListIndexes(
         string serviceName,
         RetryPolicyOptions? retryPolicy = null);
 
-    /// <summary>
-    /// Gets the full definition of a search index
-    /// </summary>
-    /// <param name="serviceName">The name of the search service</param>
-    /// <param name="indexName">The name of the search index</param>
-    /// <param name="retryPolicy">Optional retry policy for the operation</param>
-    /// <returns>The search index definition object</returns>
     Task<SearchIndexProxy?> DescribeIndex(
         string serviceName,
         string indexName,
         RetryPolicyOptions? retryPolicy = null);
 
-    /// <summary>
-    /// Queries an Azure Search index
-    /// </summary>
-    /// <param name="serviceName">The name of the search service</param>
-    /// <param name="indexName">The name of the search index</param>
-    /// <param name="searchText">The search query text</param>
-    /// <param name="retryPolicy">Optional retry policy for the operation</param>
-    /// <returns>The search results excluding large vector fields</returns>
     Task<List<JsonElement>> QueryIndex(
         string serviceName,
         string indexName,